@@ -1,39 +1,27 @@
-description = 'Ad Service'
-
-buildscript {
-    repositories {
-        mavenCentral()
-        mavenLocal()
-        maven {
-            url "https://plugins.gradle.org/m2/"
-        }
-    }
-    dependencies {
-        classpath 'com.google.protobuf:protobuf-gradle-plugin:0.8.3'
-        classpath "gradle.plugin.com.github.sherter.google-java-format:google-java-format-gradle-plugin:0.7.1"
-    }
+plugins {
+    id 'com.google.protobuf' version '0.9.3'
+    id 'com.github.sherter.google-java-format' version '0.9'
+    id 'idea'
+    id 'application'
 }
-
-apply plugin: 'idea'
-apply plugin: 'java'
-apply plugin: 'com.google.protobuf'
-apply plugin: 'com.github.sherter.google-java-format'
 
 repositories {
     mavenCentral()
     mavenLocal()
 }
 
+description = 'Ad Service'
 group = "adservice"
 version = "0.1.0-SNAPSHOT"
 
-def opencensusVersion = "0.18.0"
-def grpcVersion = "1.17.0"
-def jacksonVersion = "2.9.6"
+def grpcVersion = "1.54.1"
+def jacksonCoreVersion = "2.15.0"
+def jacksonDatabindVersion = "2.15.0"
+def protocVersion = "3.22.3"
 
 tasks.withType(JavaCompile) {
-    sourceCompatibility = '1.8'
-    targetCompatibility = '1.8'
+    sourceCompatibility = JavaVersion.VERSION_19
+    targetCompatibility = JavaVersion.VERSION_19
 }
 
 ext {
