@@ -22,18 +22,6 @@ import hipstershop.Demo.AdResponse;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.StatusRuntimeException;
-import io.opencensus.common.Duration;
-import io.opencensus.common.Scope;
-import io.opencensus.contrib.grpc.metrics.RpcViews;
-import io.opencensus.contrib.grpc.util.StatusConverter;
-import io.opencensus.exporter.stats.stackdriver.StackdriverStatsConfiguration;
-import io.opencensus.exporter.stats.stackdriver.StackdriverStatsExporter;
-import io.opencensus.exporter.trace.stackdriver.StackdriverTraceConfiguration;
-import io.opencensus.exporter.trace.stackdriver.StackdriverTraceExporter;
-import io.opencensus.trace.Span;
-import io.opencensus.trace.Tracer;
-import io.opencensus.trace.Tracing;
-import io.opencensus.trace.samplers.Samplers;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.Nullable;
 import org.apache.logging.log4j.Level;
@@ -44,7 +32,6 @@ import org.apache.logging.log4j.Logger;
 public class AdServiceClient {
 
   private static final Logger logger = LogManager.getLogger(AdServiceClient.class);
-  private static final Tracer tracer = Tracing.getTracer();
 
   private final ManagedChannel channel;
   private final hipstershop.AdServiceGrpc.AdServiceBlockingStub blockingStub;
@@ -55,7 +42,7 @@ public class AdServiceClient {
         ManagedChannelBuilder.forAddress(host, port)
             // Channels are secure by default (via SSL/TLS). For the example we disable TLS to avoid
             // needing certificates.
-            .usePlaintext(true)
+            .usePlaintext()
             .build());
   }
 
@@ -75,36 +62,24 @@ public class AdServiceClient {
     AdRequest request = AdRequest.newBuilder().addContextKeys(contextKey).build();
     AdResponse response;
 
-    Span span =
-        tracer
-            .spanBuilder("AdsClient")
-            .setRecordEvents(true)
-            .setSampler(Samplers.alwaysSample())
-            .startSpan();
-    try (Scope scope = tracer.withSpan(span)) {
-      tracer.getCurrentSpan().addAnnotation("Getting Ads");
+    try {
       response = blockingStub.getAds(request);
-      tracer.getCurrentSpan().addAnnotation("Received response from Ads Service.");
     } catch (StatusRuntimeException e) {
-      tracer.getCurrentSpan().setStatus(StatusConverter.fromGrpcStatus(e.getStatus()));
       logger.log(Level.WARN, "RPC failed: " + e.getStatus());
       return;
-    } finally {
-      span.end();
-    }
+    } 
     for (Ad ads : response.getAdsList()) {
       logger.info("Ads: " + ads.getText());
     }
   }
 
-  private static int getPortOrDefaultFromArgs(String[] args, int index, int defaultPort) {
-    int portNumber = defaultPort;
-    if (index < args.length) {
+  private static int getPortOrDefaultFromArgs(String[] args) {
+    int portNumber = 9555;
+    if (2 < args.length) {
       try {
-        portNumber = Integer.parseInt(args[index]);
+        portNumber = Integer.parseInt(args[2]);
       } catch (NumberFormatException e) {
-        logger.warn(
-            String.format("Port %s is invalid, use default port %d.", args[index], defaultPort));
+        logger.warn(String.format("Port %s is invalid, use default port %d.", args[2], 9555));
       }
     }
     return portNumber;
@@ -127,42 +102,7 @@ public class AdServiceClient {
     // Add final keyword to pass checkStyle.
     final String contextKeys = getStringOrDefaultFromArgs(args, 0, "camera");
     final String host = getStringOrDefaultFromArgs(args, 1, "localhost");
-    final int serverPort = getPortOrDefaultFromArgs(args, 2, 9555);
-
-    // Registers all RPC views.
-    RpcViews.registerAllGrpcViews();
-
-    // Registers Stackdriver exporters.
-    long sleepTime = 10; /* seconds */
-    int maxAttempts = 3;
-
-    for (int i = 0; i < maxAttempts; i++) {
-      try {
-        StackdriverTraceExporter.createAndRegister(StackdriverTraceConfiguration.builder().build());
-        StackdriverStatsExporter.createAndRegister(
-            StackdriverStatsConfiguration.builder()
-                .setExportInterval(Duration.create(15, 0))
-                .build());
-      } catch (Exception e) {
-        if (i == (maxAttempts - 1)) {
-          logger.log(
-              Level.WARN,
-              "Failed to register Stackdriver Exporter."
-                  + " Tracing and Stats data will not reported to Stackdriver. Error message: "
-                  + e.toString());
-        } else {
-          logger.info("Attempt to register Stackdriver Exporter in " + sleepTime + " seconds");
-          try {
-            Thread.sleep(TimeUnit.SECONDS.toMillis(sleepTime));
-          } catch (Exception se) {
-            logger.log(Level.WARN, "Exception while sleeping" + e.toString());
-          }
-        }
-      }
-    }
-
-    // Register Prometheus exporters and export metrics to a Prometheus HTTPServer.
-    // PrometheusStatsCollector.createAndRegister();
+    final int serverPort = getPortOrDefaultFromArgs(args);
 
     AdServiceClient client = new AdServiceClient(host, serverPort);
     try {
