@@ -13,14 +13,17 @@
 // limitations under the License.
 
 const cardValidator = require('simple-card-validator');
-const uuid = require('uuid/v4');
+const { v4: uuidv4 } = require('uuid');
 const pino = require('pino');
 
 const logger = pino({
   name: 'paymentservice-charge',
   messageKey: 'message',
-  changeLevelName: 'severity',
-  useLevelLabels: true
+  formatters: {
+    level (logLevelString, logLevelNum) {
+      return { severity: logLevelString }
+    }
+  }
 });
 
 
@@ -53,7 +56,7 @@ class ExpiredCreditCard extends CreditCardError {
  * Verifies the credit card number and (pretend) charges the card.
  *
  * @param {*} request
- * @return transaction_id - a random uuid v4.
+ * @return transaction_id - a random uuid.
  */
 module.exports = function charge (request) {
   const { amount, credit_card: creditCard } = request;
@@ -79,5 +82,5 @@ module.exports = function charge (request) {
   logger.info(`Transaction processed: ${cardType} ending ${cardNumber.substr(-4)} \
     Amount: ${amount.currency_code}${amount.units}.${amount.nanos}`);
 
-  return { transaction_id: uuid() };
+  return { transaction_id: uuidv4() };
 };
