@@ -16,19 +16,44 @@
 
 'use strict';
 
-require('@google-cloud/profiler').start({
-  serviceContext: {
-    service: 'paymentservice',
-    version: '1.0.0'
-  }
-});
-require('@google-cloud/trace-agent').start();
-require('@google-cloud/debug-agent').start({
-  serviceContext: {
-    service: 'paymentservice',
-    version: 'VERSION'
-  }
-});
+
+if(process.env.DISABLE_PROFILER) {
+  console.log("Profiler disabled.")
+}
+else {
+  console.log("Profiler enabled.")
+  require('@google-cloud/profiler').start({
+    serviceContext: {
+      service: 'paymentservice',
+      version: '1.0.0'
+    }
+  });
+}
+
+
+if(process.env.ENABLE_TRACING == "1") {
+  console.log("Tracing enabled.")
+  const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node');
+  const { SimpleSpanProcessor } = require('@opentelemetry/sdk-trace-base');
+  const { GrpcInstrumentation } = require('@opentelemetry/instrumentation-grpc');
+  const { registerInstrumentations } = require('@opentelemetry/instrumentation');
+  const { OTLPTraceExporter } = require("@opentelemetry/exporter-otlp-grpc");
+
+  const provider = new NodeTracerProvider();
+  
+  const collectorUrl = process.env.COLLECTOR_SERVICE_ADDR
+
+  provider.addSpanProcessor(new SimpleSpanProcessor(new OTLPTraceExporter({url: collectorUrl})));
+  provider.register();
+
+  registerInstrumentations({
+    instrumentations: [new GrpcInstrumentation()]
+  });
+}
+else {
+  console.log("Tracing disabled.")
+}
+
 
 const path = require('path');
 const HipsterShopServer = require('./server');
