@@ -32,7 +32,7 @@ func (q Quote) String() string {
 
 // CreateQuoteFromCount takes a number of items and returns a Price struct.
 func CreateQuoteFromCount(count int) Quote {
-	return CreateQuoteFromFloat(quoteByCountFloat(count))
+	return CreateQuoteFromFloat(8.99)
 }
 
 // CreateQuoteFromFloat takes a price represented as a float and creates a Price struct.
@@ -42,14 +42,4 @@ func CreateQuoteFromFloat(value float64) Quote {
 		uint32(units),
 		uint32(math.Trunc(fraction * 100)),
 	}
-}
-
-// quoteByCountFloat takes a number of items and generates a price quote represented as a float.
-func quoteByCountFloat(count int) float64 {
-	if count == 0 {
-		return 0
-	}
-	count64 := float64(count)
-	var p = 1 + (count64 * 0.2)
-	return count64 + math.Pow(3, p)
 }
